first step

+++ backend/src/db/migrations/20240706110241_create_categories_table.js
@@ -13,10 +13,13 @@ exports.up = function(knex) {
  * @param { import("knex").Knex } knex
  * @returns { Promise<void> }
  */
-exports.down = function(knex) {
-    return knex.schema.table('products', function(table) {
-        table.dropColumn('category_id');
-      }).then(() => {
-        return knex.schema.dropTable('categories');
-      });
+exports.down = async function(knex) {
+  const hasColumn = await knex.schema.hasColumn('products', 'category_id');
+  if (hasColumn) {
+    await knex.schema.alterTable('products', (table) => {
+      table.dropColumn('category_id');
+    });
+  }
+
+  await knex.schema.dropTableIfExists('categories');
 };
